REFACTOR: Eliminate tagged union of all possible analysis types.
It created unnecessary centralization without helping us do anything.
