+++ packages/frontend/src/analysis/document.ts
@@ -2,7 +2,7 @@ import type { ExternRef, LiveDoc } from "../api";
 import type { LiveDiagramDocument } from "../diagram";
 import type { LiveModelDocument } from "../model";
 import { type Notebook, newNotebook } from "../notebook";
-import type { ModelAnalysis } from "./types";
+import type { Analysis } from "./types";
 
 /** Common base type for all analysis documents. */
 export type BaseAnalysisDocument = {
@@ -14,14 +14,18 @@ export type BaseAnalysisDocument = {
     /** Reference to the document that the analysis is of. */
     analysisOf: ExternRef;
 
-    /** Content of the analysis. */
-    notebook: Notebook<unknown>;
+    /** Content of the analysis.
+
+    Because each analysis comes with its own content type and Solid component,
+    we do not bother to enumerate all possible analyses in a tagged union.
+    This means that analysis content type is `unknown`.
+     */
+    notebook: Notebook<Analysis<unknown>>;
 };
 
 /** A document defining an analysis of a model. */
 export type ModelAnalysisDocument = BaseAnalysisDocument & {
     analysisOf: { taxon: "model" };
-    notebook: Notebook<ModelAnalysis>;
 };
 
 /** A document defining an analysis of a diagram. */
